// apps/rokbattles-platform/src/lib/types/statistics.ts
export interface Statistics {
  battle: number;
  battle_npc: number;
  duelbattle2: number;
  barcanyonkillboss: number;
}

export interface StatisticsBattleAggregationDb {
  battle: number;
  battle_npc: number;
}

export interface StatisticsBarCanyonAggregationDb {
  barcanyonkillboss: number;
}